Fix unauthenticated redirection not working with ui-router 1.0
Both `unauthenticatedRedirector` and `unauthenticatedRedirectPath` have no
effect when using ui-router 1.0

This is caused by fix introduced in 02a8549 which aimed to resolve TransitionRejection
error (see: https://github.com/auth0/angular-jwt/pull/138#issuecomment-287735509)

Because of that, client is always being redirected to state defined
in `loginPath` config option (which BTW doesn't appear to be used
anywhere in the codebase or mentioned in documentation) and
`unauthenticatedRedirector` is never invoked.

The solution is to abort current transition, wait until its processing
is finished and then invoke redirector function. This can be achieved by
either chaining off the transition promise or adding onError hook.

Invoking redirector function is delayed just in case there are some problems
with using `$location` during transition, however it looks like
it is currently working fine when using ui-router 1.0.3.

TransitionRejection error still occurs, but this is correct behaviour
and it does not cause any further implications for redirector function.

Additionally, `onBefore` transition hook should be used instead of
`onStart`, so that the transition does not even begin when the client is
unauthenticated.